// frontend/src/api/generated/DashboardWidgets.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AccountWidget } from "./AccountWidget";
import type { AuditRow } from "./AuditRow";
import type { MyOrg } from "./MyOrg";
import type { NotificationsWidget } from "./NotificationsWidget";
import type { SecurityWidget } from "./SecurityWidget";

export type DashboardWidgets = { account: AccountWidget, security: SecurityWidget, notifications: NotificationsWidget, activity: Array<AuditRow>, organizations: Array<MyOrg>, };